requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.1
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
oauthlib==3.2.2
pillow==10.3.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.4.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
